# KiemTraThichUng.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KiemTraThichUng API",
    "description": "API documentation for Adaptive Testing System",
    "version": "v1"
  },
  "paths": {
    "/api/AdaptiveTest/BatDauKiemTra": {
      "post": {
        "tags": [
          "AdaptiveTest"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatDauKiemTra.BatDauKiemTraCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BatDauKiemTra.BatDauKiemTraCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BatDauKiemTra.BatDauKiemTraCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.BatDauKiemTra.BatDauKiemTraResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.BatDauKiemTra.BatDauKiemTraResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.BatDauKiemTra.BatDauKiemTraResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/AdaptiveTest/GetCauHoiTiepTheo": {
      "post": {
        "tags": [
          "AdaptiveTest"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayCauHoiTiepTheo.LayCauHoiTiepTheoCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LayCauHoiTiepTheo.LayCauHoiTiepTheoCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LayCauHoiTiepTheo.LayCauHoiTiepTheoCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.LayCauHoiTiepTheo.LayCauHoiTiepTheoResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.LayCauHoiTiepTheo.LayCauHoiTiepTheoResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.LayCauHoiTiepTheo.LayCauHoiTiepTheoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/AdaptiveTest/NopCauTraLoi": {
      "post": {
        "tags": [
          "AdaptiveTest"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NopCauTraLoi.NopCauTraLoiCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NopCauTraLoi.NopCauTraLoiCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NopCauTraLoi.NopCauTraLoiCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.NopCauTraLoi.NopCauTraLoiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.NopCauTraLoi.NopCauTraLoiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.NopCauTraLoi.NopCauTraLoiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/AdaptiveTest/KetThucKiemTra": {
      "post": {
        "tags": [
          "AdaptiveTest"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KetThucKiemTra.KetThucKiemTraCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/KetThucKiemTra.KetThucKiemTraCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/KetThucKiemTra.KetThucKiemTraCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.KetThucKiemTra.KetThucKiemTraResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.KetThucKiemTra.KetThucKiemTraResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.KetThucKiemTra.KetThucKiemTraResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/AdaptiveTest/LayDuLieuKiemTra": {
      "post": {
        "tags": [
          "AdaptiveTest"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayDuLieuKiemTra.LayDuLieuKiemTraQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LayDuLieuKiemTra.LayDuLieuKiemTraQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LayDuLieuKiemTra.LayDuLieuKiemTraQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.AdaptiveTest.DuLieuKiemTraDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.AdaptiveTest.DuLieuKiemTraDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.AdaptiveTest.DuLieuKiemTraDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/DM_CauTrucBCHs/GetListByIdParent": {
      "post": {
        "tags": [
          "DM_CauTrucBCHs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetListByIdParent.GetListByIdParentQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetListByIdParent.GetListByIdParentQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetListByIdParent.GetListByIdParentQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.PagedResult.DM_CauTrucBCHs.CauTrucItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.PagedResult.DM_CauTrucBCHs.CauTrucItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.PagedResult.DM_CauTrucBCHs.CauTrucItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/DM_CauTrucBCHs": {
      "post": {
        "tags": [
          "DM_CauTrucBCHs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCautruc.CreateCauTrucCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCautruc.CreateCauTrucCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCautruc.CreateCauTrucCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.CreateCautruc.CreateCauTrucResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.CreateCautruc.CreateCauTrucResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.CreateCautruc.CreateCauTrucResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "DM_CauTrucBCHs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCautruc.UpdateCauTrucCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCautruc.UpdateCauTrucCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCautruc.UpdateCauTrucCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.UpdateCautruc.UpdateCauTrucResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.UpdateCautruc.UpdateCauTrucResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.UpdateCautruc.UpdateCauTrucResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "DM_CauTrucBCHs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCauTruc.DeleteCauTrucCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCauTruc.DeleteCauTrucCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCauTruc.DeleteCauTrucCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.DeleteCauTruc.DeleteCauTrucResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.DeleteCauTruc.DeleteCauTrucResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse.DeleteCauTruc.DeleteCauTrucResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdaptiveTest.CauTraLoisMucTieuDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idCauHoi": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "noiDung": {
            "type": "string",
            "nullable": true
          },
          "isVeTrai": {
            "type": "boolean",
            "nullable": true
          },
          "isKhongDao": {
            "type": "boolean",
            "nullable": true
          },
          "viTriGachChan": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "stt": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ApiResponse.AdaptiveTest.DuLieuKiemTraDto": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/AdaptiveTest.DuLieuKiemTraDto"
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ApiResponse.BatDauKiemTra.BatDauKiemTraResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/BatDauKiemTra.BatDauKiemTraResponse"
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ApiResponse.CreateCautruc.CreateCauTrucResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/CreateCautruc.CreateCauTrucResponse"
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ApiResponse.DeleteCauTruc.DeleteCauTrucResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/DeleteCauTruc.DeleteCauTrucResponse"
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ApiResponse.KetThucKiemTra.KetThucKiemTraResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/KetThucKiemTra.KetThucKiemTraResponse"
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ApiResponse.LayCauHoiTiepTheo.LayCauHoiTiepTheoResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/LayCauHoiTiepTheo.LayCauHoiTiepTheoResponse"
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ApiResponse.NopCauTraLoi.NopCauTraLoiResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/NopCauTraLoi.NopCauTraLoiResponse"
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ApiResponse.PagedResult.DM_CauTrucBCHs.CauTrucItemResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/PagedResult.DM_CauTrucBCHs.CauTrucItemResponse"
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ApiResponse.UpdateCautruc.UpdateCauTrucResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/UpdateCautruc.UpdateCauTrucResponse"
          },
          "warningMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Responses.ApiMessage"
            },
            "nullable": true,
            "readOnly": true
          },
          "isOk": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "BatDauKiemTra.BatDauKiemTraCommand": {
        "type": "object",
        "properties": {
          "idCauHinhDeKiemTra": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "BatDauKiemTra.BatDauKiemTraResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idNguoiDung": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idCauHinhDeKiemTra": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "thoiGianBatDau": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "thoiGianKetThuc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "diemSo": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "isDat": {
            "type": "boolean",
            "nullable": true
          },
          "mucNangLuc": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maMucNangLuc": {
            "type": "string",
            "nullable": true
          },
          "diemNangLuc": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "saiSoUocLuong": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "idCauHoiHienTai": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "trangThai": {
            "$ref": "#/components/schemas/Enums.TrangThaiKiemTra"
          }
        },
        "additionalProperties": false
      },
...
      "CreateCautruc.CreateCauTrucCommand": {
        "type": "object",
        "properties": {
          "tenCauTruc": {
            "type": "string",
            "nullable": true
          },
          "maCauTruc": {
            "type": "string",
            "nullable": true
          },
          "idBoCauHoi": {
            "type": "integer",
            "format": "int32"
          },
          "idParent": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ghiChu": {
            "type": "string",
            "nullable": true
          },
          "stt": {
            "type": "integer",
            "format": "int32"
          },
          "isVisible": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "DeleteCauTruc.DeleteCauTrucCommand": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DeleteCauTruc.DeleteCauTrucResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DM_CauTrucBCHs.CauTrucItemResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetListByIdParent.GetListByIdParentQuery": {
        "type": "object",
        "properties": {
          "idBoCauHoi": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idParent": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "keyword": {
            "type": "string",
            "nullable": true
          },
          "sortCol": {
            "type": "string",
            "nullable": true
          },
          "isAsc": {
            "type": "boolean",
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "KetThucKiemTra.KetThucKiemTraCommand": {
        "type": "object",
        "additionalProperties": false
      },
      "KetThucKiemTra.KetThucKiemTraResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idNguoiDung": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idCauHinhDeKiemTra": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "doKhoMin": {
            "type": "string",
            "nullable": true
          },
          "doKhoMax": {
            "type": "string",
            "nullable": true
          },
          "mucNangLucDat": {
            "type": "string",
            "nullable": true
          },
          "thoiGianBatDau": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "thoiGianKetThuc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "diemSo": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "isDat": {
            "type": "boolean",
            "nullable": true
          },
          "mucNangLuc": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maMucNangLuc": {
            "type": "string",
            "nullable": true
          },
          "diemNangLuc": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "saiSoUocLuong": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "idCauHoiHienTai": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "trangThai": {
            "$ref": "#/components/schemas/Enums.TrangThaiKiemTra"
          }
        },
        "additionalProperties": false
      },
      "LayCauHoiTiepTheo.LayCauHoiTiepTheoCommand": {
        "type": "object",
        "additionalProperties": false
      },
      "LayCauHoiTiepTheo.LayCauHoiTiepTheoResponse": {
        "type": "object",
        "properties": {
          "isFinished": {
            "type": "boolean"
          },
          "reason": {
            "$ref": "#/components/schemas/ValueObjects.StopReason"
          },
          "cauHoi": {
            "$ref": "#/components/schemas/AdaptiveTest.CauHoiMucTieuDto"
          }
        },
        "additionalProperties": false
      },
...
      "LayDuLieuKiemTra.LayDuLieuKiemTraQuery": {
        "type": "object",
        "properties": {
          "idKetQuaKiemTra": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "NopCauTraLoi.NopCauTraLoiCommand": {
        "type": "object",
        "properties": {
          "dapAnNguoiDung": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdaptiveTest.CauTraLoisMucTieuDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NopCauTraLoi.NopCauTraLoiResponse": {
        "type": "object",
        "properties": {
          "thetaBefore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "thetaAfter": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "standardError": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "answerResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdaptiveTest.KetQuaDanhGiaDapAn"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Responses.ApiMessage": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateCautruc.UpdateCauTrucCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tenCauTruc": {
            "type": "string",
            "nullable": true
          },
          "maCauTruc": {
            "type": "string",
            "nullable": true
          },
          "ghiChu": {
            "type": "string",
            "nullable": true
          },
          "stt": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isVisible": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Format: Bearer {your token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}